fix(ft8): 重构为标准 FT8 编解码，实现与 WSJT-X/FT8CN 互通
修复原 FT8 实现自洽但非标准的问题：
- LDPC: (174,87) 自创码 -> 标准 (174,91) 码 + 置信传播软判决解码
- 消息打包: 网格/呼号哈希/自由文本/消息类型全部按标准算法
- CRC-14: 按 ft8_lib 逐位算法 + 77位零扩展
- 调制: 简单正弦 -> 标准 GFSK (BT=2.0)
- 解码: 硬判决位翻转 -> 多候选同步搜索 + 软似然提取 + BP
- 录音: 采样率回退+重采样, 按15s时隙对齐解码

新增 Ft8Protocol.kt 标准协议层、单元测试及真实录音互通验证。

+++ app/src/main/java/com/example/radioarealocator/data/ft8/Ft8Models.kt
@@ -36,9 +36,9 @@ enum class Ft8Band(
 @Parcelize
 data class Ft8DecodedMessage(
     val utcTime: Long,
-    val snr: Int, // 信噪比 dB
+    val snr: Float, // 信噪比 dB
     val dt: Float, // 时间偏移 秒
-    val freqHz: Int, // 音频频率
+    val freqHz: Float, // 音频频率
     val raw77: Long, // 77-bit 消息原文
     val callsign: String,
     val grid: String?,
